time series

+++ internal/core/models/point.go
@@ -3,9 +3,9 @@ package models
 import "fmt"
 
 type Point struct {
-	X, Y int
+	X, Y float64
 }
 
 func (p *Point) String() string {
-	return fmt.Sprintf("(%d, %d)", p.X, p.Y)
+	return fmt.Sprintf("(%f, %f)", p.X, p.Y)
 }
